Fix sparseindices (#742)
* accept size_t for Vec

* Prepare a test case for "SparseIndices too large" core dump

* fix python docstring deps

* fix compilation errors

* limit sdf hashtable size

* another fix to hashtable

* log2(0) is undefined

* fix formatting

* Revert "fix compilation errors"

This reverts commit adfb86c18fb8a3f44f138f8b66d583b3c1c81193.

* fix size_t stuff

* remove processOverlap setting in test

* fix hashtable

* vertex number limit

* address comments

+++ src/manifold/src/shared.h
@@ -212,7 +212,7 @@ struct ReindexEdge {
   VecView<const TmpEdge> edges;
   SparseIndices& indices;
 
-  void operator()(int i) {
+  void operator()(size_t i) {
     int& edge = indices.Get(i, inverted);
     edge = edges[edge].halfedgeIdx;
   }
